updated copy (#120)
Updated the copy on the Home and 404 pages based on the final text
provided in the issue.

Didn’t update the Footer or Coming Soon page since those components are
still in development. We can either open a new issue or update the
existing one to track those copy changes once the components are ready.

Related to #115