Scrape next meeting date (#118)
# Summary of Changes

1. Adds a script that scrapes the Boston liquor licensing board web page
for the next voting meeting date and writes it to a JSON file.
2. The database header component reads the next meeting date from the
JSON file.
3. Create a new Github Action workflow that:
    1. Checks out the repo
    2. Installs node
3. Installs axios, cheerio, and fs packages (all others in
package-lock.json are unnecessary for scraping/writing)
    4. Runs the script
    5. Adds and commits changes to the repo
 4. Repo is scheduled to run at midnight every Sunday.

# Notes & questions for reviewers

1. The workflow currently runs when a commit has been pushed to my
branch, and the changes will be committed to my branch. This is so that
reviewers can observe the workflow running and associated changes, and
will be changed before the PR is merged.
2. One scrape a week shouldn't be too much of a load for either our
Github account or the liquor license board website, but it's easy to
change that to every two weeks.
3. What username and email should we use for commits made by this
workflow? I currently have "Github Actions" and
"hello@codeforboston.org".

---------

Co-authored-by: GitHub Actions <hello@codeforboston.org>