Next meeting date create pr (#134)
# Description

Currently, the Github Action to update the Next Meeting Date attempts to
commit and push updates to the main branch. However that branch is
protected and making an exception for the GHA and only the GHA is
impractical, so we will create pull requests to make the update instead.

## Notes

1. The created PR will not have any reviewers or assignees, since
hard-coding that isn't a good idea
2. Since the action runs once a week, if the PR isn't merged or closed
in that week and a new run updates the next meeting date again, the open
PR will be updated.
3. If the PR source branch is not deleted after the PR is closed or
merged, the next run will delete that branch and create a new one.

## Other changes
Compiling the Next Meeting Date script generates a JS file that we don't
want to check in, so all JS files in that directory will be .gitignored.