Dynamic page titles (#177)
# Description

These changes will dynamically update the title displayed in the browser
tab when the user navigates to different routes and when they change
language. The `lang` attribute on the top level `<html>` element will
also be updated when the user changes language. Therefore this also
fixes issue #139.

# Detailed explanation

According to TanStack Router [documentation
](https://tanstack.com/router/latest/docs/framework/react/guide/document-head-management),
the preferred way to update `<head>` properties like `<title>` and
`<meta>` is to put in a `head` property when creating a creating a
route:

```
export const Route = createRootRoute({
  component: RootComponent,
  notFoundComponent: NotFound,
  head: () => ({
    meta: [
      {
        title: `${intl.formatMessage({ id: "notFound.pageTitle" })} | ${intl.formatMessage({ id: "home.pageTitle" })}`,
      },
    ],
  }),
});
```

Since this way requires calling the `useIntl` hook outside of a
component, we have to create a separate `intl` singleton which may cause
performance issues. In addition, since the `title` is set outside of a
component, it will not be updated when the user changes language.
Currently there is no way to update router contexts in response to
language changes.

However, if we put a `<HeaderContent />` element before the `<Outlet />`
in the root component, any `<title />` elements in components will be
hoisted up to the `<head />`, and thus we can set the page title in any
component by putting a `<title />` element inside. Since it is inside a
component, it will be rerendered and updated when the language changes.

From now on, whenever we add a new route/page, we will need to add a
`<title />` element to its main component to show the page title in the
browser tab.